Refactor folders and naming convention
This is a step to work on the direction to fix #116:

- It moves all business logic components outside components/blocks to its corresponding /components/pages/<name>/components
- It uses PascalCase for `organizations`, `elections`, `transactions`, and `blocks`
- Ir rename from process to elections
