+++ components/pages/organizations/list/OrganizationList.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode, useState } from 'react';
 
-import { OrganizationsFilter, IFilterEntity } from '../components/organizations-filter';
+import { IFilterEntity, OrganizationsFilter } from '../components/OrganizationsFilter';
 import { InlineTitleChildrenContainer } from '@components/pages/app/page-templates/list-page';
 import {
   FilteredPaginatedList,
@@ -9,7 +9,7 @@ import {
 import { useOrganizationList } from '@hooks/use-voconi-sdk';
 import { IChainOrganizationResponse } from '@vocdoni/sdk';
 import { OrganizationProvider } from '@vocdoni/chakra-components';
-import { OrganizationCard } from '@components/blocks/card/entity-card';
+import { OrganizationCard } from '@components/pages/organizations/components/OrganizationCard';
 
 interface IDashboardProcessListProps {
   loading?: boolean;
